Clarified some labels
When I first wanted to do sorting it wasn't clear if 'disc' and 'track' really represented the disc number and the track number. For instance was 'track' the track title? It wasn't clear - from context I could assume that it wasn't since there is also a 'title' label but by changing to track number and disc number this clears up that confusion outright.